refactor: 유틸리티 클래스 패키지 정리
