+++ app/src/main/java/woowacourse/shopping/view/product/detail/ProductDetailEventHandler.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.view.product.detail
 
 import woowacourse.shopping.domain.model.Product
-import woowacourse.shopping.view.common.QuantityControlEventHandler
+import woowacourse.shopping.view.util.QuantityControlEventHandler
 
 interface ProductDetailEventHandler : QuantityControlEventHandler<Product> {
     fun onAddToCartClick()
