+++ app/src/main/java/woowacourse/shopping/view/product/detail/ProductDetailViewModel.kt
@@ -7,8 +7,8 @@ import woowacourse.shopping.domain.model.Product
 import woowacourse.shopping.domain.model.RecentProduct
 import woowacourse.shopping.domain.repository.CartProductRepository
 import woowacourse.shopping.domain.repository.RecentProductRepository
-import woowacourse.shopping.view.common.MutableSingleLiveData
-import woowacourse.shopping.view.common.SingleLiveData
+import woowacourse.shopping.view.util.MutableSingleLiveData
+import woowacourse.shopping.view.util.SingleLiveData
 
 class ProductDetailViewModel(
     val product: Product,
